+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+- Renamed the `cloudsmith token` command to `cloudsmith login` (token still works).
+
 ## [0.16.0] - 2019-09-18
 
 - Added support for --content-type when uploading Raw packages.